# api/openapi.yaml
openapi: 3.1.0
info:
  title: PaNasMs prototype API
  version: 0.2.0
paths:
...
  /api/v1/cooling:
    get:
      operationId: getCooling
      security:
      - session: []
      responses: &id001
        '200':
          description: Cooling settings and last observed controller state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoolingState'
        default:
          description: Failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: setCooling
      security:
      - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoolingSettings'
      responses: *id001
  /api/v1/profile:
    get:
      operationId: getProfile
      security:
      - session: []
      responses:
        '200':
          description: Own Linux profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        default: &id002
          description: Failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      operationId: updateProfile
      security:
      - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
      responses:
        '204':
          description: Updated; password change revokes panel sessions
        default: *id002
  /api/v1/manage:
    get:
      summary: Read a management view using the authenticated administrator identity
      parameters:
      - {in: query, name: view, required: true, schema: {type: string, enum: [homes, homes-check, jobs, storage-options, raid-candidates, smart, services, journal, updates, files, nfs, power, network]}}
      - {in: query, name: target, schema: {type: string}}
      responses:
        '200': {description: Typed view result or explicit error; never shell output}
        '403': {description: Administrator required}
    post:
      summary: Preview, enqueue or cancel an operation, or clear finished job history
      parameters:
      - {in: query, name: view, required: true, schema: {type: string, enum: [plan, run, cancel, clear-history]}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                id: {type: string, description: Client idempotency identifier}
                action: {type: string, description: Allowlisted domain action; arbitrary commands are rejected}
                params: {type: object, additionalProperties: true}
                fingerprint: {type: string, description: Fresh preview state digest}
                confirmation: {type: string, description: Exact target acknowledgement from preview}
      responses:
        '200': {description: Preview with target/details/fingerprint/confirmation, or existing idempotent job}
        '202': {description: Persistent job accepted}
        '409': {description: Conflict or task can no longer be cancelled}
...
components:
  schemas:
...
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
...
    CoolingSettings:
      type: object
      required:
      - profile
      - sampleSeconds
      - cpuProfile
      additionalProperties: false
      properties:
        profile:
          type: string
          enum:
          - quiet
          - balanced
          - performance
        sampleSeconds:
          type: integer
          minimum: 30
          maximum: 600
        cpuProfile:
          type: string
          enum:
          - quiet
          - balanced
          - performance
    CoolingState:
      type: object
      required:
      - config
      - status
      - available
      properties:
        config:
          $ref: '#/components/schemas/CoolingSettings'
        available:
          type: boolean
        status:
          type: object
          required:
          - dutyPercent
          - reason
          - disks
          - observedAt
          - profile
          properties:
            profile:
              type: string
            dutyPercent:
              type: integer
            reason:
              type: string
            observedAt:
              type: number
            disks:
              type: array
              items:
                type: object
                required:
                - path
                - state
                - temperature
                properties:
                  path:
                    type: string
                  state:
                    type: string
                    enum:
                    - active
                    - sleeping
                    - unknown
                    - unavailable
                  temperature:
                    type:
                    - number
                    - 'null'
                  device:
                    type: string
                  health:
                    type: string
                    enum:
                    - passed
                    - warning
                    - failed
                    - unknown
                  stale:
                    type: boolean
                  observedAt:
                    type:
                    - number
                    - 'null'
                  warnings:
                    type: array
                    items:
                      type: string
                  attributes:
                    type: array
                    items:
                      type: object
                      required:
                      - id
                      - name
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        raw:
                          type: object
                          properties:
                            value:
                              type: number
                            string:
                              type: string
            cpu:
              type: object
              required:
              - available
              properties:
                available:
                  type: boolean
                profile:
                  type: string
                thresholds:
                  type: array
                  items:
                    type: integer
    Profile:
      type: object
      required:
      - username
      - name
      - uid
      - home
      - groups
      - keys
      properties:
        username:
          type: string
        name:
          type: string
        uid:
          type: string
        home:
          type: string
        groups:
          type: array
          items:
            type: string
        keys:
          type: array
          items:
            type: object
            required:
            - id
            - type
            - fingerprint
            - comment
            properties:
              id:
                type: string
              type:
                type: string
              fingerprint:
                type: string
              comment:
                type: string
    ProfileUpdate:
      type: object
      required:
      - action
      additionalProperties: false
      properties:
        action:
          type: string
          enum:
          - name
          - password
          - add
          - delete
        name:
          type: string
          maxLength: 80
        currentPassword:
          type: string
          writeOnly: true
        newPassword:
          type: string
          writeOnly: true
        key:
          type: string
        id:
          type: string
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: panasms_session